+++ backend/pkg/logger/logger.go
@@ -7,6 +7,10 @@ import (
 	"github.com/HyperloopUPV-H8/h9-backend/pkg/abstraction"
 )
 
+const (
+	dataLogger abstraction.LoggerName = "data"
+)
+
 type Logger struct {
 	running    bool
 	lock       sync.Mutex