+++ backend/pkg/logger/CHANGELOG.md
@@ -1,4 +1,12 @@
 # Logger Changelog
 
-Commit 6 "logger-structure-v3"
----
+## v0.3.0
+
+- Integrate abstraction
+- "Keys" concept implementation
+
+## v0.3.1
+
+- Data sublogger implementation
+  - Add constants
+  - Add map of valueTypes